// WFX_Code/WFXWeb/WFX.Web/ClientApp/src/models/LineTargetModel.ts
export class LineTargetModel{
  userID: number;
  lineTargetID: number;
  module: string;
  section: string;
  style: string;
  line: string;
  soNo: number;
  poNo: number;
  part: string;
  color: string;
  smv:number;
  operators:number;
  helpers:number;
  shiftName:string;
  shiftHours:string;
  date:Date;
  plannedEffeciency: number;
  plannedTarget:number
  entryDate:Date
  sizeList:number

}